version bump to make this compatible with pyproject.toml
