--- Pipfile
@@ -1,14 +1,0 @@
-[[source]]
-url = "https://pypi.org/simple"
-verify_ssl = true
-name = "pypi"
-
-[packages]
-numpy = "*"
-
-[dev-packages]
-pytest = "*"
-pillow = "*"
-
-[requires]
-python_version = "3.6"
